requirements.txt
# Twitter API
tweepy==4.14.0

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
nltk==3.8.1

# Machine Learning
scikit-learn==1.3.2

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6
